Add the initial 'Expr' type, test that it renders correctly.
We don't want to write shaders in raw GLSL because we'd miss out on all
the type safety that Haskell offers. Instead, we'll wrap our GLSL
expressions in a typed wrapper, which we can use to guide the user
towards correct operations.

This adds the basic 'Expr' wrapper, plus a couple functions that allow
us to test that it works as expected.