Make the AST explicit rather than delegating to `language-glsl`
Previously, the expression type was a newtype around `language-glsl`'s
`Expr` type. This was very convenient as the compile step was... well,
unwrapping a newtype. However, it's limiting for a number of reasons.

The big one we're about to care about is that we want to annotate nodes
in the tree with type information so that we can factor out common
expressions into intermediate bindings. `language-glsl` (understandably)
doesn't carry around any sort of type information, so we need a type
that does.

It also means we can remove _some_ of the noise around the GLSL
constructors at least in the intermediate stages.