Generalise `renderExpr`, tidy up the tests
The tests are a bit cumbersome at the moment because, in order to check
the result of a GLSL computation, we have to express that result as a
`vec4`. We do this so that we can perform the computation in a fragment
shader, then render a pixel's colour to the answer. This pixel's colour
can then be read to "extract" a value from the shader.

However, this means that every test involves generating a value,
performing a computation, converting it to a `vec4`, then converting a
`V4 GLfloat` into whatever we want. It's a bit of a pain.

This PR extracts that roundtripping logic into a new class, `Roundtrip`,
and simplifies all the tests by removing the conversions in both
directions. They're certainly easier to read.