Revert to Cofree
I think it's going to make the intermediate computation step /much/
easier if we're always talking about the same type.

We return to `Cofree TypeSpecifier ExprF` as our `Expr` representation.
When we implement observable sharing, this means we get the DeRef
structure, `CofreeF TypeSpecifier ExprF` for free. It also means that we
can use recursion schemes to generate the AST in potentially more
exciting ways (can we use a histo to avoid redundant intermediate
values?)