Separate the `shaders-test` target from `shaders` entirely
`shaders-test` already imports all the modules of `shaders` to make
ghcid work. originally, I added its dependency on `shaders` to make sure
that I wouldn't, for example, forget to add a module to both.

However, it's a bit of a nuisance: when there's an error in the code
included as part of `shaders`, I can't start a ghcid session to fix it
for `shaders-test` until I fix the dependency.

This PR removes that dependency, but also adds a CI job to make sure
that `shaders` (and future packages!) continue to build.